Update 'create_spatial_source_cadastre_wizard' wizard
