+++ asistente_ladm_col/asistente_ladm_col_plugin.py
@@ -78,7 +78,7 @@ from .gui.create_property_record_card_prc import CreatePropertyRecordCardPRCWiza
 from .gui.wizards.cadastre.rrr.create_responsibility_cadastre_wizard import CreateResponsibilityCadastreWizard
 from .gui.wizards.cadastre.rrr.create_restriction_cadastre_wizard import CreateRestrictionCadastreWizard
 from .gui.wizards.cadastre.rrr.create_right_cadastre_wizard import CreateRightCadastreWizard
-from .gui.create_spatial_source_cadastre_wizard import CreateSpatialSourceCadastreWizard
+from .gui.wizards.cadastre.source.create_spatial_source_cadastre_wizard import CreateSpatialSourceCadastreWizard
 from .gui.create_parcel_valuation_wizard import CreateParcelValuationWizard
 from .gui.create_horizontal_property_valuation_wizard import CreateHorizontalPropertyValuationWizard
 from .gui.create_common_equipment_valuation_wizard import CreateCommonEquipmentValuationWizard
