+++ asistente_ladm_col/gui/wizards/cadastre/rrr/create_responsibility_cadastre_wizard.py
@@ -261,7 +261,6 @@ class CreateResponsibilityCadastreWizard(QWizard, WIZARD_UI):
             self.exec_form(layer)
 
     def exec_form(self, layer):
-
         try:
             # Disconnect signal to prevent add features
             layer.featureAdded.disconnect()
