+++ asistente_ladm_col/gui/wizards/cadastre/rrr/create_right_cadastre_wizard.py
@@ -262,7 +262,6 @@ class CreateRightCadastreWizard(QWizard, WIZARD_UI):
             self.exec_form(layer)
 
     def exec_form(self, layer):
-
         try:
             # Disconnect signal to prevent add features
             layer.featureAdded.disconnect()
