+++ asistente_ladm_col/gui/wizards/cadastre/source/create_administrative_source_cadastre_wizard.py
@@ -185,7 +185,6 @@ class CreateAdministrativeSourceCadastreWizard(QWizard, WIZARD_UI):
             self.exec_form(layer)
 
     def exec_form(self, layer):
-
         try:
             # Disconnect signal to prevent add features
             layer.featureAdded.disconnect()
