+++ asistente_ladm_col/gui/wizards/cadastre/spatial_unit/associate_extaddress_cadastre_wizard.py
@@ -215,15 +215,9 @@ class AssociateExtAddressWizard(QWizard, WIZARD_UI):
         except:
             pass
 
-        layers = [self._layers[EXTADDRESS_TABLE]['layer'],
-                  self._layers[OID_TABLE]['layer'],
-                  self._layers[PLOT_TABLE]['layer'],
-                  self._layers[BUILDING_TABLE]['layer'],
-                  self._layers[BUILDING_UNIT_TABLE]['layer']]
-
-        for layer in layers:
+        for layer_name in self._layers:
             try:
-                layer.willBeDeleted.disconnect(self.layer_removed)
+                self._layers[layer_name]['layer'].willBeDeleted.disconnect(self.layer_removed)
             except:
                 pass
 
@@ -512,7 +506,6 @@ class AssociateExtAddressWizard(QWizard, WIZARD_UI):
             self.exec_form(layer)
 
     def exec_form(self, layer):
-
         try:
             # Disconnect signal to prevent add features
             layer.featureAdded.disconnect()
